requirements.txt
# NVIDIA Multitalker Parakeet ASR Requirements
# =============================================

# Core PyTorch with CUDA support
torch>=2.0.0
torchaudio>=2.0.0

# NVIDIA NeMo toolkit (ASR framework)
# Install from GitHub for latest multi-speaker support
nemo_toolkit[asr] @ git+https://github.com/NVIDIA/NeMo.git

# Audio processing
soundfile>=0.12.0
resampy>=0.4.0
librosa>=0.10.0

# Configuration and utilities
omegaconf>=2.3.0
hydra-core>=1.3.0

# Transformers for tokenizers
transformers>=4.30.0

# Numerical computing
numpy>=1.24.0

# Lhotse for audio data loading (used by NeMo)
lhotse>=1.20.0

# Additional dependencies that may be needed
sentencepiece>=0.1.99
webdataset>=0.2.0
